Add Azure Linux 2.0 x86_64
Azure Linux (previously known as CBL-Mariner) is an internal Linux
distribution for Microsoft's cloud infrastructure and edge products
and services. It is an RPM-based distribution.

https://github.com/microsoft/azurelinux

Azure Linux 2.0 x86_64 packages are hosted on
https://packages.microsoft.com/cbl-mariner/2.0/prod/base/x86_64/

We have two GPG keys to consider:
- azurelinux-metadata-key
   - https://github.com/microsoft/azurelinux/blob/2.0/SPECS/mariner-repos/MICROSOFT-METADATA-GPG-KEY
   - This public key can be used to verify the repomd.xml.asc signature
     for the repomd.xml
- azurelinux-rpm-key
   - https://github.com/microsoft/azurelinux/blob/2.0/SPECS/mariner-repos/MICROSOFT-RPM-GPG-KEY
   - This public key can be used to verify the krnel RPM signature
     itself

Because two keys are used for verifying repomd.xml and the kernel rpm
package signatures, we adopted the multi-key method used for Fedora
Asahi. The azurelinux.gpg file contains both keys, and the gpg-keys
Makefile is updated to create this file from the two keys. Additionally,
the MULTI_KEYS dictionary in rpm.py is updated to include the azurelinux
mapping to its underlying keys for RPM to use for verification.

Finally, an entry for Azure Linux 2.0 x86_64 is added in the config.ini
so the kernel config data can be extracted and incorporated into the
database.

Signed-off-by: Chris Co <chrco@microsoft.com>